backend/gateway/requirements.txt
fastapi
uvicorn
httpx
pydantic[email]
python-dotenv
pyjwt
